Исправлены замечания в коде
